Add files via upload
This version is more advanced. The balance verification ws removed for simplicity and testing (will be added back). Current issue with this one is a ""Cannot read properties of undefined (reading '_bn')" Error which is due probably to the swapcreator not receiving all the necesary accounts (as per "Detailed error in swapTokens: {
  message: "Cannot read properties of undefined (reading '_bn')",
  receivedData: {
    ammId: 'T8aPD2Va7MXxHFUizs7LHfYNffRR4ionsQCy55jeWc6',
    userSource: 'GCjvYFLDBYdrUNgXf3YyPrssXjKKcinnv5oMZ9EXMfiV',
    userDestination: 'GCjvYFLDBYdrUNgXf3YyPrssXjKKcinnv5oMZ9EXMfiV'
" . Will check if need to put original naming in the service(and listener?) file.  Check with the rust swap script if structure is ok and if need to rename accounts to original names. Look in the current java if there is a reason it is not receiving the accounts from the service(or if it needs the non decoded version?) file.